spritepal/requirements.txt
# SpritePal Requirements File
# Generated based on import analysis

# Core Qt Framework
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0

# Image Processing
Pillow>=9.0.0
numpy>=1.24.0

# Testing Framework
pytest>=7.2.0
pytest-qt>=4.2.0
pytest-xvfb>=2.0.0
pytest-benchmark>=4.0.0
pytest-cov>=4.0.0

# Code Quality Tools
ruff>=0.1.0
basedpyright>=1.1.0

# Development Tools
coverage>=7.0.0

# Optional but recommended
# scipy>=1.10.0  # Removed as per search feature updates
# opencv-python>=4.7.0  # For advanced image processing if needed